+++ lib/ab/version.rb
@@ -1,3 +1,3 @@
-class Ab
+module Ab
   VERSION = '0.0.1'
 end
